Use incompatible pre-release format version
Change the format version so that this code is incompatible with
existing releases, and vice versa.  We're distributing test builds and
might change the format based on feedback and we don't want it to be
possible to corrupt volumes with the changes in these pre-release
builds.

This is a one-off patch for this feature branch.  We may work this into
the build more formally if it works out.

Signed-off-by: Zach Brown <zab@versity.com>