Avoid math issues on fast test machines.
This test checks if running multiple jobs at the same time doesn't lock
down the machine for too long. This works fine usually because most
current hardware we're using for testing takes a few seconds to complete
one unit of this test in single threaded mode. But on relatively fast
new hardware, the test unit finishes in under 1 second. The timeout
threshold calculated therefore becomes (factor * unit-time) = 0, and
this makes the test always fail on these fast machines.

To fix, use a 1s bottom time for unit-time, always. We can go to a
higher precision if that becomes relevant, or figure out something
better later.

Signed-off-by: Auke Kok <auke.kok@versity.com>