Use a private copy of filefrag from e2fsprogs.
The output format of e2fsprogs has changed several times over the
past few el releases. Each time this breaks our testing since it
relies on the output format to be consistent.

Let's just be past this for a long time and import a copy of filefrag
before the last breakage. We can now then undo any of the special
massaging of the output permanently and update our golden output
to match just the private copy of it. For historical purposes, I've
just appended the git commit postfix to the source and binary so
that it's clear which thing the tests are actually using. All the
tests are updated to make sure we our private copy.

The source code of filefrag is mostly the same except I've removed
most of the special ext2 handling in it, and imported a static copy
into the C file of the fiemap.h header. This allows me to not have
to include external files from e2fsprogs to handle the ext2 specifics.

Signed-off-by: Auke Kok <auke.kok@versity.com>