Add fault injection code (clobbers) to scoutfs utility.
This fault injection code can insert 1 single error (problem) into
an existing filesystem. The goal of this functionality is to validate
the check code's ability to repair each individual problem as needed.
This will be part of our normal testing and each implemented repair
functionality should have at least one clobber equivalend.

Each clobber is independent and can be passed an arbitrary data string
to fine tune the way that the clobber operates. This allows us to have
very complex clobbers in case we need it, but also extremely simple ones
where it's not needed, and not have hundreds of extra cmdline options.

Each clobber function should have it's own description string that
contains an explanation as to how the passed data value(s) are used or
not.

Signed-off-by: Auke Kok <auke.kok@versity.com>