Check, repair and clobber for the meta device superblock CRC.
Example check and repair code for the superblock CRC value. No
actual checking is done whether the SB is fully correct, only
the CRC is rewritten to the correct value.

The clobber for the crc in the meta device's superblock flips a
randomly picked bit in the CRC. Running the clobber again may
cause the CRC to be unclobbered.

Signed-off-by: Auke Kok <auke.kok@versity.com>