Clean up debugging output
All internal output that is for developer consumption is converted to
using the `debug()` call and newlines are stripped from the end of line
since that function already appends them.

You can now pass `-d -` to `check` and `clobber` and it will
dup(STDERR_FILENO) instead of opening a new file. This allows developers
quick access to raw debugging output on the console without having to
tail or inspect a text file at a later point.

Signed-off-by: Auke Kok <auke.kok@versity.com>