Modify exit values of `scoutfs check`
We pick the standard exit code values from `fsck` and mirror their
intentional behavior. This results in `fsck.scoutfs` can now be
trivially created by making it a wrapper around `scoutfs check`.

The goal is to create exit values that can be reasonably used to
construct automation scripts that determine the state of the filesystem
being checked: Whether something was repaired, not repaired, or no
incoherency was found.

Signed-off-by: Auke Kok <auke.kok@versity.com>