Check that filesystem isn't mounted.
A lot of bad things can happen if we attempt to repair a fs that still
is or appears to be still mounted. We should really avoid automatically
repairing a mounted filesystem. So when we get errno 16 when opening
meta_dev, hard stop.

We also do some rudimentary checks on the superblock to see if there
are mounted_clients, and halt check/repair.

The --force flag is added to reset the mounted_clients count to 0. There
is a new clobber to manipulate it artificially, and the test script uses
this clobber to inject and test the bahavior of refusing to repair
unless --force is passed.

Lastly we're making more code that does little changes to the superblock
and so we add a method to writeback the superblock changes to
global_super, which allows any fixing algorithm to just make the wanted
choices as long as check_supers() has initialized it. It will also
recalculate the CRC for the caller.

Signed-off-by: Auke Kok <auke.kok@versity.com>