Avoid integer wrapping pitfalls for (off, len) pairs.
Ideally we could make some type-agnostic macro to detect that off + len
doesn't overflow the integer type, either for unsigned or signed.

But as it stands, we can't really easily get the type's max integer
positive value if it's signed programmatically.

So I've just created loff_t and u64 variants of the positive side
overflow detection code. We can use these and they're not too
complicated to replicate for new types should we need them.

For simplicity, we assume that a negative offset always wraps. We also
assume that a zero byte OR a one byte length never wraps.

And then apply these where we do wrapping checks.

simple-release-extents test mistakenly relied on this buggy wrap code,
so it needs fixing.

The move-blocks test also got it wrong.

Signed-off-by: Auke Kok <auke.kok@versity.com>