Allow skipping of tests.
With el9, we will be skipping the format-version-forward-back test on
el9 platforms, because the test requires that we compile and run code
that has no el9 support. For that test, we will rely on older OS
versions to not fail, and it will not be skipped there.

So we have to modify our outcome checks otherwise all el9 builds will
fail testing no matter what. This is the simplest form where tests
can skip - only failing tests will return an error code.

Signed-off-by: Auke Kok <auke.kok@versity.com>