Don't fence clients with rid==0
Occasionally during the export-lookup-evict-race test we see the
following failure dmesg output when a server fences a client that has no
valid rid:

[  828.379546] sysfs: cannot create duplicate filename '/fs/scoutfs/f.b928e1.r.7b36c0/fence/0000000000000000'
...
[  828.379773] kobject_add_internal failed for 0000000000000000 with -EEXIST, don't try to register things with the same name in the same directory.
[  828.385946] scoutfs f.b928e1.r.7b36c0 error: client fence returned err -17, shutting down server

This fails the test. Fencing these clients is unwanted, but we
definitely don't want to create duplicate sysfs entries for it, either.

Don't fence clients like this, just return 0.

Signed-off-by: Auke Kok <auke.kok@versity.com>