Make `df` print out blocks free based on data, only
The current `df` output displays an estimate of free disk space
available based on the usage of both meta and data device.

Change it to purely report back on data device usage only.

There are arguments both for and against this. First, it is an obvious
change in current behavior, that may break existing scripts and usage
reporting methods.

Arguments for this are that this is easier to understand for users - the
number will now reflect whether more data can be stored, or whether data
needs to be released or not to meet availability requirements.

Signed-off-by: Auke Kok <auke.kok@versity.com>