Do not fence connections without valid greeting.
There is no reason to fence any connection that hasn't sent a valid
greeting, since they haven't progressed far enough for it to make
sense.

This prevents the server from emitting:
```
    client 127.0.0.1:12345 reconnect timed out, fencing
```

But, it doesn't actually affect the behavior much. No fencing is taking
place at this time. The server will still close all listening sockets
and hand over to the next server, and all clients will disconnect
and reconnect to the new leader.

Signed-off-by: Auke Kok <auke.kok@versity.com>